requirements.txt
h5py==3.2.1
fastapi==0.115.12
learn2learn==0.2.0
matplotlib==3.4.3
numpy==1.22.0
onnx==1.16.1
onnxruntime==1.18.1
paho-mqtt==2.1.0
Pillow==10.3.0
PyWavelets==1.1.1
scikit_learn==0.24.2
scipy==1.7.1
setuptools<81
torch==2.1.1
torchvision==0.16.1
uvicorn==0.34.0
